+++ Magic/docs/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Improved spell icons to better fit vanilla style (thanks to Ash!).
 * Added magic school icons.
+* Added tooltips.
 
 ## 0.6.2
 Released 29 October 2021 for SMAPI 3.12.5 or later. Updated by Pathoschild.